Servidor/requirements.txt
# ==============================================================================
# SMART CLASSROOM AI - Dependencies
# ==============================================================================

# ---- Core Framework ----
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.12.0,<3.0.0
pydantic-settings==2.1.0

# ---- Computer Vision & AI ----
deepface==0.0.87
opencv-python-headless==4.9.0.80
# TensorFlow: usar versión compatible con tu Python
tensorflow>=2.15.0; python_version < '3.12'
tensorflow>=2.16.0; python_version >= '3.12'
tf-keras>=2.20.0  # Requerido por TensorFlow 2.20+
pillow>=10.0.0
numpy>=1.24.0,<2.0.0

# ---- Database & Vector Search ----
supabase>=2.9.0,<2.10.0
gotrue>=2.12.0,<3.0.0
postgrest>=0.17.0,<0.18.0
realtime>=2.27.0,<3.0.0
storage3>=0.8.0,<0.9.0
supafunc>=0.6.0,<0.7.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pgvector==0.2.4

# ---- Async & HTTP ----
httpx>=0.27.0,<0.28.0
websockets>=15.0.0
aiofiles==23.2.1
python-multipart==0.0.6

# ---- Security ----
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# ---- Utils ----
python-dateutil==2.8.2
pydantic-core==2.14.6

# ---- Development (Optional) ----
pytest==7.4.3
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
